split preference options into its own file
